Add code for classifying and visualizing text using fasttext and tSNE
This includes our parametric_tsne package as a dependency.
We script the execution of creating a fasttext model with all 14 classes included, and visualizing the result with parametric_tsne.
We also have a script for filtering out a specific class from the training/test data, building a model based on only 13 classes, and visualizing the result.